moved metrics and logger in a separate folder, renammed logger to Continuum logger because it is not exactly a metric logger but more prediction logger somehow. Maybe a better name might be found
