+++ continuum/metrics/__init__.py
@@ -1,0 +1,5 @@
+# pylint: disable=C0401
+# flake8: noqa
+from continuum.metrics.continuum_logger import ContinuumLogger
+
+__all__ = ["ContinuumLogger"]
